+++ mybooking-templates/mybooking-plugin-profile-change-password.php
@@ -1,0 +1,11 @@
+<div class="row">
+  <div class="col-lg-12">
+    <form name="mybooking_change_password_form" autocomplete="off">
+      <div class="form-group">
+          <label for="mybooking_change_password_password"><?php echo esc_html_x( "Password", 'profile_change_password', 'mybooking-wp-plugin') ?></label>
+          <input type="password" name="password" class="form-control" id="mybooking_change_password_password" placeholder="<?php echo esc_html_x( "Password", 'profile_change_password', 'mybooking-wp-plugin') ?>">
+      </div>
+      <button type="submit" class="btn btn-primary"><?php echo esc_html_x( "Send", 'profile_change_password', 'mybooking-wp-plugin') ?></button>      
+    </form>
+  </div>
+</div>